pyproject.toml
[tool.poetry]
name = "config"
version = "0.1.0"
description = "Feelody backend project"
authors = ["shin <shz0130@naver.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = ">=3.13"
djangorestframework = ">=3.15.2,<4.0.0"
isort = ">=6.0.1,<7.0.0"
black = ">=25.1.0,<26.0.0"
psycopg2 = "^2.9.9"
django-dotenv = ">=1.4.2,<2.0.0"
requests = ">=2.32.3,<3.0.0"
python-dotenv = ">=1.0.1,<2.0.0"
djangorestframework-simplejwt = ">=5.5.0,<6.0.0"
pytest = ">=8.3.5,<9.0.0"
pytest-django = ">=4.10.0,<5.0.0"
ytmusicapi = ">=1.10.2,<2.0.0"
apiclient = ">=1.0.4,<2.0.0"
google-api-python-client = ">=2.165.0,<3.0.0"
google-generativeai = ">=0.8.4,<0.9.0"
drf-yasg = ">=1.21.10,<2.0.0"
responses = ">=0.25.7,<0.26.0"
gunicorn = ">=23.0.0,<24.0.0"
django-cors-headers = "^4.7.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^5.0.1"

[tool.black]
line-length = 80

[tool.isort]
profile = "black"
